+++ aegis_gym/rsl/envs/grasp_env.py
@@ -10,7 +10,6 @@ import torch as th
 import torch.nn.functional as F
 from PIL import Image
 
-from rsl_rl.env import VecEnv
 from tensordict import TensorDict
 from genesis.utils.geom import (
     transform_by_quat,
@@ -18,14 +17,15 @@ from genesis.utils.geom import (
 )
 
 from config_types.domain_randomization import DomainRandomizationCfg, CameraPoseCfg
-from .manipulator import Manipulator
+from .manipulator import GenesisManipulator
+from .base_env import BaseEnv, StepReturn, ResetReturn
 from .plotjuggler_udp import PlotJugglerUDP
 
 # Further example
 # https://github.com/isaac-sim/IsaacLab/blob/857da263c08fa78664e40ab957f996b22153d181/source/isaaclab_rl/isaaclab_rl/rsl_rl/vecenv_wrapper.py
 
 
-class GraspEnv(VecEnv):
+class GraspEnv(BaseEnv):
     def __init__(
         self,
         env_cfg: dict,
@@ -34,6 +34,7 @@ class GraspEnv(VecEnv):
         show_viewer: bool = False,
         enable_plot_juggler: bool = False,
     ) -> None:
+        super().__init__(scene=None)  # TODO(issue#128) introduce Scene abstraction
         if enable_plot_juggler:
             ip = "127.0.0.1"
             port = 9870
@@ -84,7 +85,7 @@ class GraspEnv(VecEnv):
             # env_spacing=(1.0, 1.0),
         )
 
-        self.robot.set_pd_gains()
+        self.robot.set_joints_pd_gains()
         self._attach_cameras()
 
         if self._dr_cfg.enabled:
@@ -170,7 +171,7 @@ class GraspEnv(VecEnv):
         )
 
         # == add robot ==
-        self.robot = Manipulator(
+        self.robot = GenesisManipulator(
             num_envs=self.num_envs,
             scene=self.scene,
             args=robot_cfg,
@@ -298,26 +299,21 @@ class GraspEnv(VecEnv):
         ]
 
         for cam_name, link_name, offset in cams_to_attach:
+            # TODO(issue#127) migrate this code into Scene/Manipulator layer
             link = self.robot._robot_entity.get_link(link_name)
             for cam_dict in (self._cameras, self._debug_cameras):
                 if cam_name in cam_dict:
                     cam_dict[cam_name].attach(link, offset)
                     cam_dict[cam_name].move_to_attach()
 
-    # Required by rsl_rl
-    @property
-    def unwrapped(self) -> "GraspEnv":
-        return self
-
-    # Required by rsl_rl
-    @property
-    def step_dt(self) -> float:
+    def get_policy_dt(self) -> float:
         return self.policy_dt
 
-    # Required by rsl_rl
-    @property
-    def cfg(self) -> dict:
+    def get_cfg_as_dict(self) -> dict:
         return self._cfg
+
+    def get_num_envs(self) -> int:
+        return self.num_envs
 
     def _init_reward_functions(self) -> None:
         self.reward_functions, self.episode_sums = dict(), dict()
@@ -346,8 +342,9 @@ class GraspEnv(VecEnv):
             return
         self.episode_length_buf[envs_idx] = 0
 
-        # reset robot
-        self.robot.reset(envs_idx)
+        # Reset the robot
+        self.robot.ctrl_gripper_open(envs_idx)
+        self.robot.ctrl_go_to_home(envs_idx)
 
         # reset object
         num_reset = len(envs_idx)
@@ -435,13 +432,13 @@ class GraspEnv(VecEnv):
         self.object.set_quat(object_quat)
         self.goal_pose[:] = pose
 
-    def reset(self) -> tuple[TensorDict, dict]:
+    def reset(self) -> ResetReturn:
         self.reset_buf[:] = True
         self.reset_idx(th.arange(self.num_envs, device=gs.device))
         self._log_state_to_plot_juggler()
-        return self.get_observations(), self.extras
+        return ResetReturn(self.get_observations(), self.extras)
 
-    def step(self, actions: th.Tensor) -> tuple[TensorDict, th.Tensor, th.Tensor, dict]:
+    def step(self, actions: th.Tensor) -> StepReturn:
         # Update time
         self.episode_length_buf += 1
 
@@ -453,7 +450,7 @@ class GraspEnv(VecEnv):
         actions[:, :3] *= max_lin_speed
         actions[:, 3:] *= max_ang_speed
 
-        self.robot.apply_action(actions, open_gripper=True)
+        self.robot.ctrl_apply_vel_action(actions, open_gripper=True)
         self.scene.step()
         # TODO(issue#117) redesign the visualize-cameras feature
         if self.show_cameras_gui:
@@ -477,7 +474,7 @@ class GraspEnv(VecEnv):
         # get observations and fill extras
         obs = self.get_observations()
         dones = self.reset_buf
-        return obs, reward, dones, self.extras
+        return StepReturn(obs, reward, dones, self.extras)
 
     def _get_max_speed_coeefs(self) -> tuple[float, float]:
         cfg = self._dr_cfg.max_speed
@@ -521,7 +518,7 @@ class GraspEnv(VecEnv):
 
         print(f">>> Moving to relative goal for {move_steps} steps.")
         if joints_diff is not None:
-            self.robot.apply_dof_rel_action(joints_diff)
+            self.robot.ctrl_apply_joints_diff_action(joints_diff)
         elif cart_diff is not None:
             from math import ceil
 
@@ -536,7 +533,7 @@ class GraspEnv(VecEnv):
             print(f">>> Scaled down target: {scaled_cart_diff}")
             for action_id in range(actions_num):
                 print(f">>> Applying action #{action_id + 1}")
-                self.robot.apply_action(scaled_cart_diff, open_gripper=True)
+                self.robot.ctrl_apply_vel_action(scaled_cart_diff, open_gripper=True)
                 for _ in range(steps_per_action):
                     self.scene.step()
                     self._log_state_to_plot_juggler()
@@ -564,15 +561,13 @@ class GraspEnv(VecEnv):
         return self.reset_buf.nonzero(as_tuple=True)[0]
 
     def get_observations(self) -> TensorDict:
-        ee_pos, ee_quat = (
-            self.robot.ee_pose[:, :3],
-            self.robot.ee_pose[:, 3:7],
-        )
+        tcp_pose = self.robot.get_tcp_pose()
+        tcp_pos, tcp_quat = tcp_pose[:, :3], tcp_pose[:, 3:]
         obj_pos, obj_quat = self.object.get_pos(), self.object.get_quat()
 
         obs_components = [
-            ee_pos - obj_pos,  # 3D position difference
-            ee_quat,  # current orientation (w, x, y, z)
+            tcp_pos - obj_pos,  # 3D position difference
+            tcp_quat,  # current orientation (w, x, y, z)
             obj_pos,  # goal position
             obj_quat,  # goal orientation (w, x, y, z)
         ]
@@ -692,8 +687,8 @@ class GraspEnv(VecEnv):
         self._record_step = record_step + 1
 
     def _reward_keypoints(self) -> th.Tensor:
-        ee_pos = self.robot.ee_pose[:, :3]
-        ee_quat = self.robot.ee_pose[:, 3:7]
+        tcp_pose = self.robot.get_tcp_pose()
+        tcp_pos, tcp_quat = tcp_pose[:, :3], tcp_pose[:, 3:]
         keypoints_offset = self.keypoints_offset
         object_offset = th.tensor(
             [0.0, 0.0, -0.08],
@@ -702,8 +697,8 @@ class GraspEnv(VecEnv):
         ).repeat(self.num_envs, 1)
 
         finger_pos_keypoints = self._to_world_frame(
-            ee_pos + object_offset,
-            ee_quat,
+            tcp_pos + object_offset,
+            tcp_quat,
             keypoints_offset,
         )
         object_pos_keypoints = self._to_world_frame(
@@ -755,7 +750,7 @@ class GraspEnv(VecEnv):
     def grasp_and_lift_demo(self) -> float:
         total_steps = self.max_episode_length
         grab_height = 0.08
-        goal_pose = self.robot.ee_pose.clone()
+        goal_pose = self.robot.get_tcp_pose().clone()
         goal_pose[:, 2] -= grab_height
         # lift pose (above the object)
         lift_height = 0.16
@@ -777,13 +772,13 @@ class GraspEnv(VecEnv):
 
         for i in range(total_steps):
             if i < total_steps / 5:  # go down
-                self.robot.go_to_goal(goal_pose, open_gripper=True)
+                self.robot.ctrl_go_to_goal(goal_pose, open_gripper=True)
             elif i < total_steps * 2 / 5:  # grasping
-                self.robot.go_to_goal(goal_pose, open_gripper=False)
+                self.robot.ctrl_go_to_goal(goal_pose, open_gripper=False)
             elif i < total_steps * 3 / 5:  # lifting
-                self.robot.go_to_goal(lift_pose, open_gripper=False)
+                self.robot.ctrl_go_to_goal(lift_pose, open_gripper=False)
             elif i < total_steps * 4 / 5:  # final
-                self.robot.go_to_goal(final_pose, open_gripper=False)
+                self.robot.ctrl_go_to_goal(final_pose, open_gripper=False)
                 obj_pos = self.object.get_pos()
                 target_pos = final_pose[:, :3]
 
@@ -792,7 +787,7 @@ class GraspEnv(VecEnv):
 
                 hold_counter[in_target] += 1
             else:  # reset
-                self.robot.go_to_goal(reset_pose, open_gripper=True)
+                self.robot.ctrl_go_to_goal(reset_pose, open_gripper=True)
             self.scene.step()
 
         success = hold_counter >= hold_steps_required
@@ -805,7 +800,7 @@ class GraspEnv(VecEnv):
         if not cfg.enabled:
             return
 
-        n_dofs = self.robot.n_dofs
+        n_dofs = self.robot.get_n_dofs()
         kp_scale = (
             1.0
             + (th.rand(self.num_envs, n_dofs, device=self.device) * 2.0 - 1.0)
@@ -817,7 +812,7 @@ class GraspEnv(VecEnv):
             * cfg.kv_noise
         )
 
-        self.robot.set_pd_gains(kp_gain=kp_scale, kv_gain=kv_scale)
+        self.robot.set_joints_pd_gains(kp_gain=kp_scale, kv_gain=kv_scale)
 
     def _cache_camera_base_offsets(self) -> None:
         self._dr_cam_base_offsets = {}
@@ -1039,6 +1034,7 @@ class GraspEnv(VecEnv):
             perturbed_offset = (base_offset @ perturb).astype(np.float32)
 
             try:
+                # TODO(issue#127) change API to expose robot_entity
                 link = self.robot._robot_entity.get_link(
                     self._cameras_link_names[cam_name]
                 )
@@ -1146,6 +1142,7 @@ class GraspEnv(VecEnv):
             return
 
         data = {}
+        # TODO(issue#128) change api to expose the robot entity
         robot = self.robot._robot_entity
         for name in self._joint_names:
             j = robot.get_joint(name=name)