+++ aegis_gym/rsl/envs/scene/__init__.py
@@ -1,0 +1,5 @@
+from .base_scene import BaseScene
+
+__all__ = [
+    "BaseScene",
+]